+++ derive/src/derive/std/fmt/mod.rs
@@ -33,10 +33,17 @@ pub(crate) mod write;
 
 fmt_impl!(debug, Debug, ["Debug", "fmt::Debug"]);
 fmt_impl!(display, Display, ["Display", "fmt::Display"]);
+#[cfg(feature = "fmt")]
 fmt_impl!(binary, Binary, ["fmt::Binary"]);
+#[cfg(feature = "fmt")]
 fmt_impl!(lower_exp, LowerExp, ["fmt::LowerExp"]);
+#[cfg(feature = "fmt")]
 fmt_impl!(lower_hex, LowerHex, ["fmt::LowerHex"]);
+#[cfg(feature = "fmt")]
 fmt_impl!(octal, Octal, ["fmt::Octal"]);
+#[cfg(feature = "fmt")]
 fmt_impl!(pointer, Pointer, ["fmt::Pointer"]);
+#[cfg(feature = "fmt")]
 fmt_impl!(upper_exp, UpperExp, ["fmt::UpperExp"]);
+#[cfg(feature = "fmt")]
 fmt_impl!(upper_hex, UpperHex, ["fmt::UpperHex"]);